requirements.txt
asgiref==3.11.1
Django==5.2.13
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
PyJWT==2.12.1
psycopg[binary]==3.2.10
drf-spectacular==0.29.0
sqlparse==0.5.5
tzdata==2026.1
